project1_rag/requirements.txt
langchain==0.1.20
langchain-openai==0.1.6
langchain-community==0.0.38
openai==1.30.1
faiss-cpu==1.8.0
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
python-dotenv==1.0.1
pypdf==4.2.0
pydantic==2.7.1
ragas==0.1.9
httpx==0.27.0
